requirements.txt
Flask>=2.0
openai>=1.0.0
mysql-connector-python>=9.3.0
bcrypt>=4.0.0
python-dotenv==0.19.0
cryptography>=44.0.3
gunicorn>=21.2.0
